Implement authentication layout for login and registration pages with form validation
